requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
APScheduler==3.11.0
attrs==25.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
firebase-admin==6.2.0
Flask==2.0.1
frozenlist==1.6.0
geographiclib==1.52
geopy==2.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.3.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.4.3
packaging==25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyfcm==2.0.8
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==0.19.0
requests==2.26.0
rsa==4.9.1
tqdm==4.67.1
twilio==9.6.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==2.3.0
yarl==1.20.0
